+++ src/modules/financial/components/AnimationContainer.vue
