+++ src/modules/financial/components/CompleteFlowVisualization.vue
