+++ src/modules/financial/components/FinancialForm.vue
@@ -81,8 +81,8 @@
 import { computed, reactive, ref } from 'vue'
 import type { FinancialFormData } from '../types/financial.types'
 import { useRouter } from 'vue-router'
-import { useRegistrationStore } from '../stores/registration.store'
-import OtpModal from './OtpModal.vue'
+import { useRegistrationStore } from '../../registration/stores/registration.store'
+import OtpModal from '../../registration/components/OtpModal.vue'
 import ValidationLoader from './ValidationLoader.vue'
 import AnimationContainer from './AnimationContainer.vue'
 import CompleteFlowVisualization from './CompleteFlowVisualization.vue'
